+++ tokio/src/time/timer/stack.rs
@@ -1,5 +1,5 @@
-use crate::timer::timer::Entry;
-use crate::timer::wheel;
+use crate::time::timer::Entry;
+use crate::time::wheel;
 
 use std::ptr;
 use std::sync::Arc;
